infra/requirements.txt
aws-cdk-lib>=2.161.1,<3.0
constructs>=10.0.0,<11.0.0
